% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel_helpers.R
\name{create_header}
\alias{create_header}
\title{create_header}
\usage{
create_header(wb, sheet, name, row, col, col_span = 1)
}
\arguments{
\item{wb}{The excel workbook}

\item{sheet}{The sheet of the workbook}

\item{name}{The name of the new header}

\item{row}{The row of the cell}

\item{col}{The column of the cell}

\item{col_span}{The number of columns the header covers}
}
\description{
This function writes a header that can cover multiple columns.
To create a row of headers like this, use create_header_row
}
