% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel_helpers.R
\name{create_header_row}
\alias{create_header_row}
\title{create_header_row}
\usage{
create_header_row(wb, sheet, x, startRow, startCol)
}
\arguments{
\item{wb}{The excel workbook}

\item{sheet}{The sheet of the workbook}

\item{x}{A list of the form `list(list("col_name", 1), list("col_name_2", 3))`. Where the first element in
each list is the column name and the second element is the number of columns that the header should span over.}

\item{startRow}{The row this function will write to}

\item{startCol}{The column where this row of names will start}
}
\description{
This function takes in a list to execute create_header
multiple times on a single row
}
