% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_lossrun.R
\name{create_lossrun}
\alias{create_lossrun}
\title{create_lossrun}
\usage{
create_lossrun(val_date, trans_ = trans)
}
\arguments{
\item{val_date}{date the valuation date of the loss run.  Claim values from `trans`
will be values as of the `val_date`}

\item{trans}{data frame of claims transactions}
}
\value{
data frame of claims (1 claim per row) valued as of the `val_date`
}
\description{
view losses as of a specific date
}
