% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_helpers.R
\name{data_helpers}
\alias{data_helpers}
\alias{get_states}
\alias{get_accident_years}
\title{Data Utility Helpers}
\usage{
get_states(dat)

get_accident_years(dat)
}
\value{
The return value, if any, from executing the utility.
}
\description{
Various data manipulation functions.
}
