% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_excel_helpers.R
\name{exhibit_header_right}
\alias{exhibit_header_right}
\title{exhibit_header_right}
\usage{
exhibit_header_right(wb, sheet, start_row, start_col, x)
}
\arguments{
\item{wb}{The excel workbook}

\item{sheet}{The sheet of the workbook}

\item{start_row}{The first row of the headers}

\item{start_col}{The column of the headers}

\item{x}{A vector of strings to be written to the workbook}
}
\description{
This function writes a column of text aligned to the right
anywhere in an excel workbook
}
