% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_helpers.R
\name{show_names}
\alias{show_names}
\title{show_names}
\usage{
show_names(nms)
}
\arguments{
\item{nms}{character vector of names from the data}
}
\description{
show_names
}
\examples{
show_names(names(trans))

}
